Cleaned up some test code. Created a new exceptions folder to house all the exceptions.
