+++ topchef_client/__init__.py
@@ -2,5 +2,4 @@
 Contains a package for consuming the TopChef API
 """
 from .client import Client
-from .client import NetworkError, ValidationError
-from .client import ProcessingError
+from .exceptions import NetworkError, ValidationError, ProcessingError
